pyproject.toml
[project]
name = "mcp-mail-function"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.13.1",
    "python-dotenv>=1.1.1",
]
